Apply build scripts from previous incarnation
The Justfile and build scripts from my old version were very complicated and I was struggling a bit with them as I was seeking to add more complexity before actually having a stable release.

The UBlue image-template still has 90% of the needed functionality and is way simpler. So, I'm restarting here and will add only the necessary complexity as I go.